refactor: split 2,052-line render.js into modular page components
PHASE 4: Modular Render Split - Major maintainability improvement

Split monolithic render.js (2,052 lines) into focused, maintainable modules:

NEW MODULES (6 files):
- renderHelpers.js (171 lines) - Shared rendering utilities
  * renderPlayerTable() - Generic player table with fixtures
  * attachRiskTooltipListeners() - Risk tooltip event handlers

- transferHelpers.js (326 lines) - Transfer committee logic
  * findReplacements() - Smart replacement suggestions
  * renderProblemPlayerRow() - Problem player table rows
  * renderReplacementRow() - Replacement suggestion rows
  * window.toggleReplacements() - Expandable suggestions UI

- renderMyTeam.js (777 lines) - My Team page
  * renderMyTeamForm() - Team ID input form with auto-load
  * renderMyTeam() - Full team analysis with problem players
  * window.loadAndRenderTeam() - Team loading handler
  * window.resetMyTeam() - Clear cache handler
  * window.toggleProblemPlayers() - Problem section toggle

- renderTransferCommittee.js (242 lines) - Transfer Committee page
  * renderTransferCommittee() - Problem players + replacements

- renderDataAnalysis.js (578 lines) - Data Analysis page
  * renderDataAnalysis() - Overview & differentials tabs
  * Module-level state for filters (position, ownership, etc.)

- renderSearch.js (158 lines) - Search page
  * renderSearch() - Player search with filters
  * window.performPlayerSearch() - Search handler
  * window.filterByPosition() - Position filter handler

BENEFITS:
✅ Improved maintainability - Each page in its own module
✅ Better code organization - Shared logic properly extracted
✅ Easier debugging - Smaller, focused files
✅ Faster development - Clear separation of concerns
✅ Same functionality - All 147 tests still passing
✅ No performance impact - Dynamic imports preserve lazy loading

UPDATED:
- main.js - Updated dynamic imports to use new modules

DELETED:
- render.js - Removed 2,052-line monolith

Tests: All 147 tests passing (utils: 47, risk: 46, fixtures: 54)
